Added python-pip and virtualenv
I added python-pip and virtualenv to the to be installed packages for Ubuntu 18.04. While running the ansible playbook remote through ssh,the task to create a virtualenv failed due to the absence of setuptools. Installing python-pip and virtualenv  on the target machine solved the issue. This allows you to run an image build process from your CI/CD pipeline instead on the spot every time.